requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.1

# Apache Spark & Delta Lake
pyspark==3.5.0
delta-spark==3.0.0

# Data Warehouse & Transformation
dbt-core==1.7.4
dbt-postgres==1.7.4
sqlalchemy==2.0.23

# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.6.0
apache-airflow-providers-postgres==5.9.0

# Machine Learning - Forecasting
tensorflow==2.15.0
keras==2.15.0
prophet==1.1.5
statsmodels==0.14.1
pmdarima==2.0.4

# Machine Learning - Anomaly Detection
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2
pyod==1.1.2

# Deep Learning
torch==2.1.2
torchvision==0.16.2

# Visualization
plotly==5.18.0
folium==0.15.1
streamlit==1.29.0
dash==2.14.2
matplotlib==3.8.2
seaborn==0.13.0
geopandas==0.14.1
contextily==1.4.0

# Graph Analysis
networkx==3.2.1
python-louvain==0.16
graph-tool==2.55

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.16.0
scrapy==2.11.0
requests==2.31.0
lxml==4.9.4
httpx==0.25.2

# Geospatial
shapely==2.0.2
fiona==1.9.5
rtree==1.1.0

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1

# Data Quality & Validation
great-expectations==0.18.7
pandera==0.17.2

# API & Web Framework
fastapi==0.108.0
uvicorn==0.25.0
flask==3.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2
cloudpickle==3.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
ipywidgets==8.1.1
notebook==7.0.6

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Time Series
tslearn==0.6.3
sktime==0.25.0
